Settings Menu UI Work
Work on the Settings Menu UI, separating it into Submenus for maintainability and expansion, and starting work on making it more readable.
