+++ GPW - Space Station/Assets/Code/Scripts/UI/Menus/SettingsMenu.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 51a90abbb99970c4db25737c2883dbdd
-PrefabImporter:
+guid: 0ee69118460defe4780f39923e4043df
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
