+++ GPW - Space Station/Assets/Code/Scripts/UI/Menus/SettingsMenu/ControlsMenuUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b1c4ab53b2b8be64cac670c2a6f012bb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
