+++ GPW - Space Station/Assets/Code/Scripts/UI/Menus/SettingsMenu/DisplayMenuUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd74e99e1be147e4cbfdf57f04922f44
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
