+++ GPW - Space Station/Assets/Code/Scripts/UI/Menus/SettingsMenu/SettingsSubmenuUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04edb6525e00a464ead239bafdd66498
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
