+++ GPW - Space Station/Assets/Level/Prefabs/UI/SettingsMenu.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b3983cea358f029438d33f6fb94f0bc3
+guid: 51a90abbb99970c4db25737c2883dbdd
 PrefabImporter:
   externalObjects: {}
   userData:
